ctfs/CTF Try Out/Guild/web_guild/guild/requirements.txt
flask
flask-sqlalchemy
flask-login
Pillow